+++ src/lib.rs
@@ -11,9 +11,11 @@ async fn health_check() -> HttpResponse {
 }
 
 pub fn run(listener: TcpListener) -> Result<Server, io::Error> {
-    let server = HttpServer::new(|| App::new().route("/health_check", web::get().to(health_check)))
-        .listen(listener)?
-        .run();
+    let server = HttpServer::new(|| {
+        App::new().route("/health_check", web::get().to(health_check))
+    })
+    .listen(listener)?
+    .run();
 
     Ok(server)
 }